feat: add BarChart to personal analysis
